Elimina archivos de prueba de IR con Pixy y simplifica el código de OpenMV para la detección de objetivos.
