--- Test_IR_with_Pixy.cpp
@@ -1,11 +1,0 @@
-#include "Test_IR_with_Pixy.h"  
-
-Test_IR_with_Pixy::Test_IR_with_Pixy()
-{
-
-}
-
-Test_IR_with_Pixy::~Test_IR_with_Pixy()
-{
-
-}
