--- Test_IR_with_Pixy.h
@@ -1,16 +1,0 @@
-
-#ifndef ARDUINO_TEST_IR_WITH_PIXY_H
-#define ARDUINO_TEST_IR_WITH_PIXY_H
-#include <Arduino.h>
-
-class Test_IR_with_Pixy  
-{
-	private:
-
-	public:
-
-		Test_IR_with_Pixy();
-		~Test_IR_with_Pixy();
-
-};
-#endif
